requirements.txt
aiohttp==3.8.5
aiosignal==1.3.1
annotated-types==0.5.0
asgiref==3.7.2
async-timeout==4.0.3
attrs==23.1.0
certifi==2023.7.22
cfgv==3.4.0
charset-normalizer==3.2.0
distlib==0.3.7
Django==4.2.5
django-cors-headers==4.2.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
filelock==3.12.4
flake8==4.0.1
flake8-docstrings==1.6.0
frozenlist==1.4.0
identify==2.5.29
idna==3.4
isort==5.10.1
mccabe==0.6.1
multidict==6.0.4
nodeenv==1.8.0
openai==0.28.0
platformdirs==3.10.0
pre-commit==2.17.0
psycopg2==2.9.7
pycodestyle==2.8.0
pydantic==2.3.0
pydantic-settings==2.0.3
pydantic_core==2.6.3
pydocstyle==6.3.0
pyflakes==2.4.0
PyJWT==2.8.0
python-dotenv==1.0.0
pytz==2023.3.post1
PyYAML==6.0.1
requests==2.31.0
snowballstemmer==2.2.0
sqlparse==0.4.4
toml==0.10.2
tqdm==4.66.1
typing_extensions==4.8.0
urllib3==2.0.5
virtualenv==20.24.5
yarl==1.9.2
